% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{identity}
\alias{identity}
\title{Returns the (first) argument (safely).}
\usage{
identity(x, ...)
}
\arguments{
\item{x}{The argument to return (unmodified).}

\item{...}{Ignored.}
}
\value{
\code{x}; if \code{x} is missing, \code{NULL}.
}
\description{
Returns the (first) argument (safely).
}
\details{
Unlike \code{base::identity}, if \code{x} is missing here, \code{NULL} is returned.
}
