% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepend_class}
\alias{prepend_class}
\title{Prepend classname(s) to an object's S3 class attribute.}
\usage{
prepend_class(obj, cls)
}
\arguments{
\item{obj}{A simple (or S3) object.}

\item{cls}{A character vector of classnames (in appropriate S3 inheritance order)}
}
\value{
\code{obj} with the \code{cls} prepended to the classnames.
}
\description{
Prepend classname(s) to an object's S3 class attribute.
}
\details{
When \code{obj} is an object with reference semantics (refsem) (e.g. an {R6} object), this function will not only return the new object, but will modify \code{obj} \emph{in-place}.
This is a by-product of how \code{structure()} works and is tough to avoid with refsem objects.
The stepwise alternative (setting \code{class(obj) <- something} inside the function does the same: mutates the original object).
A future version of this function may warn when called on a refsem object (and that warning will be suppressable).
}
