% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint-package.R
\docType{package}
\name{splint-package}
\alias{splint}
\alias{splint-package}
\title{splint: Type Splints}
\description{
Splints are like casts, but weaker (though perhaps more useful).
}
\author{
\strong{Maintainer}: Murat Tasan \email{mmuurr@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
