% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint.R
\name{splint_if_missing}
\alias{splint_if_missing}
\title{Explicit missingness instruction}
\usage{
splint_if_missing(splint, val, exactly = FALSE)
}
\arguments{
\item{splint}{the splint to wrap.}

\item{val}{the value to use when the splint element is missing.}

\item{exactly}{should \code{val} be used \emph{exactly} (as-is), or should it be passed to \code{splint()}?}
}
\description{
This wraps a splint in higher-order splint structures (lists and tbls) and specifies what to do if the element defined by \code{splint} itself is missing.
If the element is missing, what value do we pass to the constituent splint?
Here we declare that value, \code{val}.
\itemize{
\item If \code{exactly} is \code{FALSE}, then \code{val} is passed to the \code{splint}.
\item If \code{exactly} is \code{TRUE}, then the \code{splint} is bypassed entirely and \code{val} is simply used in its stead.
}
}
