% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splint.R
\name{splint_simple}
\alias{splint_simple}
\title{Simple splint}
\usage{
splint_simple(f = identity, klass = character(0), force = TRUE)
}
\arguments{
\item{f}{the underlying splint function.}

\item{klass}{the classname(s) to be prepended to the splint's output.}
}
\description{
Wraps a function as a simple, atomic splint.
While the resultant splint itself is 'atomic', \code{f()} can accept whatever type argument you like.
}
\details{
When called (as a function), a simple splint will look at the single passed argument \code{x}.
\itemize{
\item If \code{x} is missing, set \code{x} to \code{NULL}.
\item If \code{x} is a missing-val sentinel (with \code{val} and \code{exactly} properties):
\itemize{
\item If \code{exactly} is \code{TRUE}, return \code{val}.
\item Else return \code{f(val)}.
}
\item Else return \code{f(x)}.
}
}
